Organize API key rules by permission domain
